requirements.txt
# =============================================================================
# MAIN REQUIREMENTS - LIGHTWEIGHT & FREQUENTLY UPDATED
# =============================================================================
# Install base requirements first: pip install -r requirements-base.txt
# Then install these: pip install -r requirements.txt
#
# WARNING: DO NOT add torch, torchvision, or torchaudio here!
# PyTorch is installed via CUDA 12.8 index in Dockerfile for RTX 5070 support

# Include heavyweight base packages
-r requirements-base.txt

# ===== WEB FRAMEWORK & API =====
fastapi==0.115.13
uvicorn[standard]==0.34.3
python-multipart==0.0.9

# Database drivers moved to requirements-base.txt for better caching

# ===== CONFIGURATION & VALIDATION =====
pydantic==2.11.7
pydantic-settings==2.3.4
python-dotenv==1.1.0

# ===== HTTP & ASYNC =====
aiohttp==3.12.13
httpx==0.28.1
aiofiles>=23.0.0
nest-asyncio==1.6.0
requests==2.32.3

# ===== UTILITIES =====
tenacity==8.2.3
tqdm==4.67.1
pyyaml==6.0.2
watchdog>=3.0.0
schedule>=1.2.0

# ===== CLI & OUTPUT =====
rich==14.0.0
typer>=0.12.0,<0.16.0

# ===== DEVELOPMENT & TESTING =====
pytest==8.4.1
pytest-asyncio>=0.21.0
black==25.1.0
flake8==7.3.0
debugpy==1.8.2

# ===== OPTIONAL DEVELOPMENT =====
# Uncomment these for development/analysis
# openpyxl==3.1.5
# ipykernel==6.29.5
# fpdf2==2.7.8